ml_model/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
google-generativeai>=0.3.0
